+++ pyporcc/utils.py
@@ -85,7 +85,7 @@ def constrain(x, lower, upper):
     MATLAB constrain
     Constrain between upper and lower limits, and do not ignore NaN
     """
-    if isinstance(x, np.float):
+    if isinstance(x, float):
         if x < lower:
             x = lower
         elif x > upper:
